Update the Release setup.

+++ copyAPIFromDevToRelease.m
@@ -1,26 +1,39 @@
 function copyAPIFromDevToRelease(NameValuePair)
-%% Copy the utility source files to the "Release > ModelingUtilityForSimscape" folder.
+% Copy the utility API source files to the "Release > ModelingUtilityForSimscape" folder.
+%
+% This function assumes that the source files are version-managed with git.
+%
 % This function copies all "+*Util" folders under the Devel folder to
 % the destination "Release > ModelingUtilityForSimscape" folder.
 % This function deletes the ModelingUtilityForSimscape folder if it already exists before copying.
 
-% Copyright 2025 The MathWorks, Inc.
+% Copyright 2025-2026 The MathWorks, Inc.
 
 arguments (Input)
-  NameValuePair.RepoTopFolder (1,1) string {mustBeFolder} = "C:\local\modutil\modeling-utility"
   NameValuePair.DryRun (1,1) logical = true
 end  % arguments
 
-assert(pwd == NameValuePair.RepoTopFolder)
+try
+  repo = gitrepo;
+catch exception
 
-devel_top_folder = fullfile(NameValuePair.RepoTopFolder, "Devel");
+  rethrow(exception)
+
+end  % try, catch
+repo_top_folder = repo.WorkingFolder;
+
+devel_top_folder = fullfile(repo_top_folder, "Devel");
 assert(isfolder(devel_top_folder))
 
-destination_folder = fullfile(NameValuePair.RepoTopFolder, "Release", "ModelingUtilityForSimscape");
-if isfolder(destination_folder)
-  rmdir(destination_folder, "s")
+destination_folder = fullfile(repo_top_folder, "Release", "ModelingUtilityForSimscape");
+if not(NameValuePair.DryRun)
+  if isfolder(destination_folder)
+
+    rmdir(destination_folder, "s")
+
+  end  % if
+  mkdir(destination_folder)
 end  % if
-mkdir(destination_folder)
 
 source_folders = matlab.buildtool.io.FileCollection.fromPaths(fullfile(devel_top_folder, "**", "+*Util*")).paths';
 target_folders = extractAfter(source_folders, "Util"+("/"|"\"));
